requirements-production.txt
# Linescope Server Production Requirements
# Optimized for performance and security

# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Production WSGI Server
gunicorn==21.2.0

# Image Processing (optimized)
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0

# Data handling
tzdata>=2024.1

# Caching and Performance
redis==5.0.1
Flask-Caching==2.1.0

# Monitoring and Logging
psutil==5.9.6

# Security
Flask-Talisman==1.1.0

# Optional: Database support (uncomment if needed)
# SQLAlchemy==2.0.23
# Flask-SQLAlchemy==3.1.1
# psycopg2-binary==2.9.9  # PostgreSQL
# PyMySQL==1.1.0  # MySQL

# Optional: Task Queue (uncomment if needed)
# celery==5.3.4
# redis==5.0.1

# Optional: Monitoring (uncomment if needed)
# sentry-sdk[flask]==1.38.0
# prometheus-flask-exporter==0.23.0

# Development and Testing (not installed in production)
# pytest==7.4.3
# pytest-flask==1.3.0
# black==23.11.0
# flake8==6.1.0